\name{LT_FTR_BIDS}
\docType{data}
\alias{LT_FTR_BIDS}
\title{Bids data from the long term FTR auctions.}
\description{
  Bid data from the long term FTR auctions from 2006 to 2010.  Shows bid
  status (CLEARED/DENIED), bid.price, bid.mw.  No participant info.

  Run \code{NEPOOL$.make.LT.FTR.auction.bids.RData() to update the data.}
}
\usage{data(LT_FTR_BIDS)}
\format{One bid per row.}
\source{NEPOOL http://www.iso-ne.com/markets/hstdata/mkt_offer_bid/lt_ftr_acutn_hist/index.html}
\keyword{datasets}

